requirements.txt
streamlit>=1.32.0
plotly>=5.19.0
pandas>=2.0.0
gspread>=6.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=5.1.0
